kb-rag-system/requirements.txt
# Core dependencies
pinecone>=5.0.0
openai>=1.0.0
python-dotenv>=1.0.0

# API Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data processing
tqdm>=4.66.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# Utilities
python-multipart>=0.0.9
tiktoken>=0.5.0